Revert "[Improve][Catalog] refactor catalog (#4540)"
This reverts commit b0a701cb83a2353da3b1f93a483c19cddd55747c.
