+++ seatunnel-connectors-v2/connector-jdbc/src/main/java/org/apache/seatunnel/connectors/seatunnel/jdbc/internal/dialect/JdbcDialect.java
@@ -17,8 +17,6 @@
 
 package org.apache.seatunnel.connectors.seatunnel.jdbc.internal.dialect;
 
-import org.apache.seatunnel.api.table.catalog.CatalogTable;
-import org.apache.seatunnel.api.table.catalog.TablePath;
 import org.apache.seatunnel.connectors.seatunnel.jdbc.config.JdbcSourceConfig;
 import org.apache.seatunnel.connectors.seatunnel.jdbc.internal.converter.JdbcRowConverter;
 
@@ -194,41 +192,5 @@ public interface JdbcDialect extends Serializable {
             Connection conn, JdbcSourceConfig jdbcSourceConfig) throws SQLException {
         PreparedStatement ps = conn.prepareStatement(jdbcSourceConfig.getQuery());
         return ps.getMetaData();
-    }
-
-    default String listDatabases() {
-        return "SHOW DATABASES;";
-    }
-
-    default String getUrlFromDatabaseName(String baseUrl, String databaseName, String suffix) {
-        return baseUrl + databaseName + suffix;
-    }
-
-    default String createDatabaseSql(String databaseName) {
-        return String.format("CREATE DATABASE IF NOT EXISTS %s;", quoteIdentifier(databaseName));
-    }
-
-    default String dropDatabaseSql(String databaseName) {
-        return String.format("DROP DATABASE IF EXISTS %s;", quoteIdentifier(databaseName));
-    }
-
-    default String getTableName(ResultSet rs) throws SQLException {
-        return rs.getString(1);
-    }
-
-    default String getTableName(TablePath tablePath) {
-        return tablePath.getTableName();
-    }
-
-    default String listTableSql(String databaseName) {
-        return "SHOW TABLES;";
-    }
-
-    default String getDropTableSql(String tableName) {
-        return String.format("DROP TABLE %s IF EXIST;", tableName);
-    }
-
-    default String createTableSql(TablePath tablePath, CatalogTable catalogTable) {
-        return "";
     }
 }
