+++ seatunnel-connectors-v2/connector-starrocks/src/main/java/org/apache/seatunnel/connectors/seatunnel/starrocks/catalog/StarRocksCatalog.java
@@ -17,9 +17,13 @@
 
 package org.apache.seatunnel.connectors.seatunnel.starrocks.catalog;
 
+import org.apache.seatunnel.api.table.catalog.Catalog;
 import org.apache.seatunnel.api.table.catalog.CatalogTable;
+import org.apache.seatunnel.api.table.catalog.PhysicalColumn;
 import org.apache.seatunnel.api.table.catalog.PrimaryKey;
+import org.apache.seatunnel.api.table.catalog.TableIdentifier;
 import org.apache.seatunnel.api.table.catalog.TablePath;
+import org.apache.seatunnel.api.table.catalog.TableSchema;
 import org.apache.seatunnel.api.table.catalog.exception.CatalogException;
 import org.apache.seatunnel.api.table.catalog.exception.DatabaseAlreadyExistException;
 import org.apache.seatunnel.api.table.catalog.exception.DatabaseNotExistException;
@@ -32,7 +36,6 @@ import org.apache.seatunnel.api.table.type.PrimitiveByteArrayType;
 import org.apache.seatunnel.api.table.type.SeaTunnelDataType;
 import org.apache.seatunnel.common.exception.CommonErrorCode;
 import org.apache.seatunnel.common.utils.JdbcUrlUtil;
-import org.apache.seatunnel.connectors.seatunnel.jdbc.catalog.AbstractJdbcCatalog;
 import org.apache.seatunnel.connectors.seatunnel.starrocks.exception.StarRocksConnectorException;
 
 import org.apache.commons.lang3.StringUtils;
@@ -44,18 +47,22 @@ import com.mysql.cj.MysqlType;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
+import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 
 import static com.google.common.base.Preconditions.checkArgument;
 
-public class StarRocksCatalog extends AbstractJdbcCatalog {
+public class StarRocksCatalog implements Catalog {
 
     protected final String catalogName;
     protected String defaultDatabase = "information_schema";
@@ -64,6 +71,8 @@ public class StarRocksCatalog extends AbstractJdbcCatalog {
     protected final String baseUrl;
     protected String defaultUrl;
     private final JdbcUrlUtil.UrlInfo urlInfo;
+
+    private static final Set<String> SYS_DATABASES = new HashSet<>();
     private static final Logger LOG = LoggerFactory.getLogger(StarRocksCatalog.class);
 
     static {
@@ -72,8 +81,6 @@ public class StarRocksCatalog extends AbstractJdbcCatalog {
     }
 
     public StarRocksCatalog(String catalogName, String username, String pwd, String defaultUrl) {
-
-        super(catalogName, username, pwd, JdbcUrlUtil.getUrlInfo(defaultUrl));
 
         checkArgument(StringUtils.isNotBlank(username));
         checkArgument(StringUtils.isNotBlank(pwd));
@@ -90,9 +97,104 @@ public class StarRocksCatalog extends AbstractJdbcCatalog {
     }
 
     @Override
-    protected boolean createTableInternal(TablePath tablePath, CatalogTable table)
-            throws CatalogException {
-        throw new UnsupportedOperationException("Unsupported create table");
+    public List<String> listDatabases() throws CatalogException {
+        try (Connection conn = DriverManager.getConnection(defaultUrl, username, pwd)) {
+
+            PreparedStatement ps = conn.prepareStatement("SHOW DATABASES;");
+
+            List<String> databases = new ArrayList<>();
+            ResultSet rs = ps.executeQuery();
+
+            while (rs.next()) {
+                String databaseName = rs.getString(1);
+                if (!SYS_DATABASES.contains(databaseName)) {
+                    databases.add(rs.getString(1));
+                }
+            }
+
+            return databases;
+        } catch (Exception e) {
+            throw new CatalogException(
+                    String.format("Failed listing database in catalog %s", this.catalogName), e);
+        }
+    }
+
+    @Override
+    public List<String> listTables(String databaseName)
+            throws CatalogException, DatabaseNotExistException {
+        if (!databaseExists(databaseName)) {
+            throw new DatabaseNotExistException(this.catalogName, databaseName);
+        }
+
+        try (Connection conn =
+                DriverManager.getConnection(
+                        urlInfo.getUrlWithDatabase(databaseName), username, pwd)) {
+            PreparedStatement ps = conn.prepareStatement("SHOW TABLES;");
+
+            ResultSet rs = ps.executeQuery();
+
+            List<String> tables = new ArrayList<>();
+
+            while (rs.next()) {
+                tables.add(rs.getString(1));
+            }
+
+            return tables;
+        } catch (Exception e) {
+            throw new CatalogException(
+                    String.format("Failed listing database in catalog %s", catalogName), e);
+        }
+    }
+
+    @Override
+    public CatalogTable getTable(TablePath tablePath)
+            throws CatalogException, TableNotExistException {
+        if (!tableExists(tablePath)) {
+            throw new TableNotExistException(catalogName, tablePath);
+        }
+
+        String dbUrl = urlInfo.getUrlWithDatabase(tablePath.getDatabaseName());
+        try (Connection conn = DriverManager.getConnection(dbUrl, username, pwd)) {
+            Optional<PrimaryKey> primaryKey =
+                    getPrimaryKey(tablePath.getDatabaseName(), tablePath.getTableName());
+
+            PreparedStatement ps =
+                    conn.prepareStatement(
+                            String.format(
+                                    "SELECT * FROM %s WHERE 1 = 0;",
+                                    tablePath.getFullNameWithQuoted()));
+
+            ResultSetMetaData tableMetaData = ps.getMetaData();
+
+            TableSchema.Builder builder = TableSchema.builder();
+            for (int i = 1; i <= tableMetaData.getColumnCount(); i++) {
+                SeaTunnelDataType<?> type = fromJdbcType(tableMetaData, i);
+                // TODO add default value and test it
+                builder.column(
+                        PhysicalColumn.of(
+                                tableMetaData.getColumnName(i),
+                                type,
+                                tableMetaData.getColumnDisplaySize(i),
+                                tableMetaData.isNullable(i) == ResultSetMetaData.columnNullable,
+                                null,
+                                tableMetaData.getColumnLabel(i)));
+            }
+
+            primaryKey.ifPresent(builder::primaryKey);
+
+            TableIdentifier tableIdentifier =
+                    TableIdentifier.of(
+                            catalogName, tablePath.getDatabaseName(), tablePath.getTableName());
+            return CatalogTable.of(
+                    tableIdentifier,
+                    builder.build(),
+                    buildConnectorOptions(tablePath),
+                    Collections.emptyList(),
+                    "");
+        } catch (Exception e) {
+            throw new CatalogException(
+                    String.format("Failed getting table %s", tablePath.getFullName()), e);
+        }
     }
 
     @Override
@@ -104,11 +206,6 @@ public class StarRocksCatalog extends AbstractJdbcCatalog {
     @Override
     public void dropTable(TablePath tablePath, boolean ignoreIfNotExists)
             throws TableNotExistException, CatalogException {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    protected boolean dropTableInternal(TablePath tablePath) throws CatalogException {
         throw new UnsupportedOperationException();
     }
 
@@ -133,11 +230,6 @@ public class StarRocksCatalog extends AbstractJdbcCatalog {
     }
 
     @Override
-    protected boolean createDatabaseInternal(String databaseName) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
     public void dropDatabase(TablePath tablePath, boolean ignoreIfNotExists)
             throws DatabaseNotExistException, CatalogException {
         try (Connection conn = DriverManager.getConnection(defaultUrl, username, pwd)) {
@@ -154,14 +246,8 @@ public class StarRocksCatalog extends AbstractJdbcCatalog {
         }
     }
 
-    @Override
-    protected boolean dropDatabaseInternal(String databaseName) throws CatalogException {
-        throw new UnsupportedOperationException();
-    }
-
     /** @see com.mysql.cj.MysqlType */
-    @Override
-    public SeaTunnelDataType<?> fromJdbcType(ResultSetMetaData metadata, int colIndex)
+    private SeaTunnelDataType<?> fromJdbcType(ResultSetMetaData metadata, int colIndex)
             throws SQLException {
         MysqlType starrocksType = MysqlType.getByName(metadata.getColumnTypeName(colIndex));
         switch (starrocksType) {
@@ -229,8 +315,7 @@ public class StarRocksCatalog extends AbstractJdbcCatalog {
     }
 
     @SuppressWarnings("MagicNumber")
-    @Override
-    public Map<String, String> buildConnectorOptions(TablePath tablePath) {
+    private Map<String, String> buildConnectorOptions(TablePath tablePath) {
         Map<String, String> options = new HashMap<>(8);
         options.put("connector", "starrocks");
         options.put("url", baseUrl + tablePath.getDatabaseName());
@@ -284,6 +369,29 @@ public class StarRocksCatalog extends AbstractJdbcCatalog {
         return res;
     }
 
+    @Override
+    public String getDefaultDatabase() {
+        return defaultDatabase;
+    }
+
+    @Override
+    public void open() throws CatalogException {
+        try (Connection conn = DriverManager.getConnection(defaultUrl, username, pwd)) {
+            // test connection, fail early if we cannot connect to database
+            conn.getCatalog();
+        } catch (SQLException e) {
+            throw new CatalogException(
+                    String.format("Failed connecting to %s via JDBC.", defaultUrl), e);
+        }
+
+        LOG.info("Catalog {} established connection to {}", catalogName, defaultUrl);
+    }
+
+    @Override
+    public void close() throws CatalogException {
+        LOG.info("Catalog {} closing", catalogName);
+    }
+
     protected Optional<PrimaryKey> getPrimaryKey(String schema, String table) throws SQLException {
 
         List<String> pkFields = new ArrayList<>();
@@ -305,5 +413,22 @@ public class StarRocksCatalog extends AbstractJdbcCatalog {
             return Optional.of(PrimaryKey.of(pkName, pkFields));
         }
         return Optional.empty();
+    }
+
+    @Override
+    public boolean databaseExists(String databaseName) throws CatalogException {
+        checkArgument(StringUtils.isNotBlank(databaseName));
+
+        return listDatabases().contains(databaseName);
+    }
+
+    @Override
+    public boolean tableExists(TablePath tablePath) throws CatalogException {
+        try {
+            return databaseExists(tablePath.getDatabaseName())
+                    && listTables(tablePath.getDatabaseName()).contains(tablePath.getTableName());
+        } catch (DatabaseNotExistException e) {
+            return false;
+        }
     }
 }
